2017.7.3.1
Release bigupdate

+++ controller/API_Controller.php
@@ -1,10 +1,41 @@
 <?php // File tiêu chuẩn
+/*
+Các field cơ bản vd: tên của bảng = DB_PREFIX.db_print_card
+id							= autonumber
+db_print_card_id			= khóa chính khi liên kết bảng
+db_print_card_name			= khóa phụ
+db_print_card_desc			= mô tả
+db_print_card_create_date	= timelapse thời gian tạo
+db_print_card_create_by		= Người tạo ra row
+db_print_card_mod_date		= timelapse thời gian sửa cuối cùng
+db_print_card_mod_by		= Người sửa cuối cùng
+db_print_card_trash_flag	= đánh dấu đã chuyển vào thùng rác
+db_print_card_trash_date	= ngày chuyển vào thùng rác
+db_print_card_trash_by		= người chuyển vào thùng rác
+db_print_card_restore_date	= ngày chuyển vào thùng rác
+db_print_card_restore_by	= người chuyển vào thùng rác
++ ... các field nội dung khác
+
+Các action trừ get ra thì đều lưu lại bảng action history
+Các field của db_action
+id -> autonumber
+db_action_date		= timelapse thời gian thực hiện action
+db_action_by		= action đc thực hiện bởi ai
+db_action_type		= tên action
+db_action_success	= kết quả action true: thành công, false: thất bại
+db_action_content	= nội dung của action table.id=id
+*/
 class API_Controller extends MVC_Controller
 {
+	private $tableName = NULL;
+
 	public function __construct() {
         parent::__construct();
+
 		if (!isset($_GET['t']))
 			Error(500);
+
+		$this->tableName = $_GET['t'];
 
 		$this->model->Load('API');
     }
@@ -12,13 +43,13 @@ class API_Controller extends MVC_Controller
     public function getAction()	{ // done
     	$sql = [
     		'select'=> [],
-    		'from'	=> DB_PREFIX.$_GET['t'],
+    		'from'	=> DB_PREFIX.$this->tableName,
     	];
 
 		$data = [
 			'response' => [
 				'success'	=> true,
-				'data'		=> $this->model->API->ExecuteQuery($sql),
+				'data'		=> $this->model->API->ExecuteQuery($sql, $this->uid),
 			],
 		];
 
@@ -28,9 +59,12 @@ class API_Controller extends MVC_Controller
 	public function newAction() { // done
 		$sql = [
 			'insert' => [
-				'id'	=> 'temp'.rand(1,100),
+				'id' => 'temp'.rand(1,100),
+				$this->tableName.'_create_date' => date(DB_DATE_FORMAT), // auto create date
+				$this->tableName.'_create_by'	  => $this->uid,
+				$this->tableName.'_trash_flag'  => false,
 			],
-			'into'	=> $_GET['t']
+			'into' => $this->tableName,
 		];
 
         $result = $this->model->API->InsertRow($sql['into'], $sql['insert']);
@@ -51,7 +85,7 @@ class API_Controller extends MVC_Controller
 					'success' => false,
 	    			'message' => 'New record is not available',
 	    			'data' => [
-	    				'id' => $insert_data['id'],
+	    				'id' => $sql['insert']['id'],
 	    			],
 	    		]
 	    	];
@@ -67,11 +101,14 @@ class API_Controller extends MVC_Controller
 		$sql = [
 			'update' => [
 				$_GET['key'] => $_GET['value'],
+				$this->tableName.'_mod_date' => date(DB_DATE_FORMAT),
+				$this->tableName.'_mod_by'   => $this->uid,
 			],
-			'table'	=> $_GET['t'],
-			'where'	=> [
+			'table' => $this->tableName,
+			'where' => [
 				[
-					'id'	=> $_GET['id'],
+					'id' => $_GET['id'],
+					$this->tableName.'_trash_flag' => false,
 				],
 			],
 		];
@@ -109,17 +146,20 @@ class API_Controller extends MVC_Controller
 
 		$sql = [
 			'update' => [
-				str_replace('enum_', '', str_replace('db_', '', $_GET['t'])).'_trash_flag' => true,
+				$this->tableName.'_trash_flag' => true,
+				$this->tableName.'_trash_date' => date(DB_DATE_FORMAT),
+				$this->tableName.'_trash_by'   => $this->uid,
 			],
-			'table'	=> $_GET['t'],
-			'where'	=> [
+			'table' => $this->tableName,
+			'where' => [
 				[
-					'id'	=> $_GET['id'],
+					'id' => $_GET['id'],
+					$this->tableName.'_trash_flag' => false,
 				],
 			],
 		];
 
-        $result = $this->model->API->UpdateRow($sql['table'], $sql['update'], $sql['where'][0]);
+        $result = $this->model->API->RemoveRow($sql['table'], $sql['update'], $sql['where'][0]);
 
 		if ($result) {
     		$data = [
@@ -144,20 +184,64 @@ class API_Controller extends MVC_Controller
     	$this->view->Load('json', $data);
 	}
 
+	public function restoreAction() { // done
+		if (!isset($_GET['id']))
+			Error(503);
+
+		$sql = [
+			'update' => [
+				$this->tableName.'_trash_flag' => false,
+				$this->tableName.'_restore_date' => date(DB_DATE_FORMAT),
+				$this->tableName.'_restore_by'   => $this->uid,
+			],
+			'table' => $this->tableName,
+			'where' => [
+				[
+					'id' => $_GET['id'],
+					$this->tableName.'_trash_flag' => true,
+				],
+			],
+		];
+
+        $result = $this->model->API->RestoreRow($sql['table'], $sql['update'], $sql['where'][0]);
+
+		if ($result) {
+    		$data = [
+	    		'response' => [
+					'success' => true,
+	    			'message' => 'Restore record id="'.$_GET['id'].'" completed',
+	    			'data' => [
+	    				'id' 	=> $_GET['id'],
+	    			]
+	    		]
+	    	];
+    	} else {
+    		$data = [
+	    		'response' => [
+					'success' => false,
+	    			'message' => 'Record id="'.$_GET['id'].'" not found',
+	    			'data' => [],
+	    		]
+	    	];
+    	}
+
+    	$this->view->Load('json', $data);
+	}
+
 	public function deleteAction() { // done
 		if (!isset($_GET['id']))
 			Error(503);
 
 		$sql = [
-			'delete'=> $_GET['t'],
+			'delete'=> $this->tableName,
 			'where'	=> [
 				[
-					'id'	=> $_GET['id'],
+					'id' => $_GET['id'],
 				],
 			],
 		];
 
-		$result = $this->model->API->DeleteRow($sql['delete'], $sql['where'][0]);
+		$result = $this->model->API->DeleteRow($sql['delete'], $sql['where'][0], $this->uid);
 
     	if ($result) {
     		$data = [
